# routes/seller.js
const express = require("express");
const router = express.Router();
const User = require("../models/User");
const Product = require("../models/Product");
const Sale = require("../models/Sale");
const SellerStock = require("../models/SellerStock");
const { authenticate, isSeller } = require("../middleware/auth");

// All sellerId routes require authentication and sellerId role
router.use(authenticate);
router.use(isSeller);

// Get sellerId's assigned products
router.get("/products", async (req, res) => {
  try {
    const products = await SellerProduct.find({
      sellerId: req.user._id,
      isActive: true,
    })
      .populate("product")
      .select("product assignAt");
    res.json({ products: products });
  } catch (error) {
    res.status(500).json({ error: error.message });
  }
});

// Get seller's stock inventory
router.get("/stocks", async (req, res) => {
  try {
    const sellerStocks = await SellerStock.findBySeller(req.user._id);

    // Calculate total stock value
    const totalStockValue = sellerStocks.reduce((total, stock) => {
      return total + stock.quantity * (stock.product?.costPrice || 0);
    }, 0);

    const totalProducts = sellerStocks.length;
    const totalQuantity = sellerStocks.reduce(
      (total, stock) => total + stock.quantity,
      0,
    );

    res.json({
      stocks: sellerStocks,
      summary: {
        totalProducts,
        totalQuantity,
        totalStockValue,
      },
    });
  } catch (error) {
    res.status(500).json({ error: error.message });
  }
});

// Get stock for a specific product
router.get("/stocks/product/:productId", async (req, res) => {
  try {
    const { productId } = req.params;

    const stock = await SellerStock.findBySellerAndProduct(
      req.user._id,
      productId,
    );

    if (!stock) {
      return res
        .status(404)
        .json({ error: "Stock not found for this product" });
    }

    await stock.populate("product", "name sku price costPrice image");

    res.json({ stock });
  } catch (error) {
    res.status(500).json({ error: error.message });
  }
});

// Get sellerId's sales
router.get("/sales", async (req, res) => {
  try {
    const { startDate, endDate } = req.query;
    const query = { sellerId: req.user._id };

    if (startDate && endDate) {
      query.timestamp = {
        $gte: new Date(startDate),
        $lte: new Date(endDate),
      };
    }

    const sales = await Sale.find(query)
      .populate("product", "name price image")
      .sort({ timestamp: -1 });

    res.json({ sales });
  } catch (error) {
    res.status(500).json({ error: error.message });
  }
});

// Get sellerId's reports
router.get("/reports", async (req, res) => {
  try {
    const { year, month } = req.query;
    const startDate = new Date(
      year || new Date().getFullYear(),
      (month || new Date().getMonth()) - 1,
      1,
    );
    const endDate = new Date(
      year || new Date().getFullYear(),
      month || new Date().getMonth(),
      0,
      23,
      59,
      59,
    );

    const sales = await Sale.find({
      sellerId: req.user._id,
      timestamp: { $gte: startDate, $lte: endDate },
    })
      .populate("product", "name price")
      .sort({ timestamp: -1 });

    const totalSales = sales.length;
    const totalRevenue = sales.reduce((sum, sale) => sum + sale.totalAmount, 0);
    const totalQuantity = sales.reduce((sum, sale) => sum + sale.quantity, 0);

    res.json({
      period: { startDate, endDate },
      summary: {
        totalSales,
        totalRevenue,
        totalQuantity,
      },
      sales,
    });
  } catch (error) {
    res.status(500).json({ error: error.message });
  }
});

module.exports = router;