# swagger/generated-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Telegram Hisobchi Bot API",
    "version": "2.0.0",
    "description": "API documentation for Telegram Hisobchi Bot Backend - Advanced Sales Management System with SellerStock Management",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "TelegramAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-telegram-id",
        "description": "Telegram user ID (REQUIRED for all authenticated endpoints)"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "User ID"
          },
          "telegramId": {
            "type": "string",
            "description": "Telegram user ID"
          },
          "username": {
            "type": "string",
            "description": "Username"
          },
          "firstName": {
            "type": "string",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "description": "Last name"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "seller"],
            "description": "User role"
          },
          "assignedProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "description": "Products assigned to seller"
          },
          "isActive": {
            "type": "boolean",
            "description": "User active status"
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Category ID"
          },
          "name": {
            "type": "string",
            "description": "Category name"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Category creation date"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Product ID"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "description": "Product description"
          },
          "price": {
            "type": "number",
            "description": "Selling price"
          },
          "costPrice": {
            "type": "number",
            "description": "Cost price"
          },
          "category": {
            "type": "object",
            "ref": "#/components/schemas/Category",
            "description": "Product category"
          },
          "count": {
            "type": "number",
            "description": "Warehouse stock count"
          },
          "image": {
            "type": "string",
            "description": "Product image URL"
          },
          "sku": {
            "type": "string",
            "description": "Stock Keeping Unit"
          },
          "color": {
            "type": "string",
            "description": "Product color"
          },
          "isActive": {
            "type": "boolean",
            "description": "Product active status"
          },
          "assignedSellers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Sellers assigned to this product"
          }
        }
      },
      "SellerStock": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "SellerStock ID"
          },
          "seller": {
            "$ref": "#/components/schemas/User"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "quantity": {
            "type": "number",
            "description": "Stock quantity with seller",
            "minimum": 0
          },
          "lastTransferDate": {
            "type": "string",
            "format": "date-time",
            "description": "Date of last transfer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation date"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update date"
          }
        }
      },
      "Sale": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Sale ID"
          },
          "seller": {
            "$ref": "#/components/schemas/User"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity sold"
          },
          "price": {
            "type": "number",
            "description": "Price per unit"
          },
          "totalAmount": {
            "type": "number",
            "description": "Total sale amount"
          },
          "customerName": {
            "type": "string",
            "description": "Customer name"
          },
          "customerPhone": {
            "type": "string",
            "description": "Customer phone number"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Sale timestamp"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "TelegramAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/sellers/{sellerId}/stocks": {
      "get": {
        "tags": ["Admin - Stock Management"],
        "summary": "Get stocks for specific seller",
        "description": "Returns all stock records for a specific seller",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sellerId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Seller ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Seller stocks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sellerStocks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SellerStock"
                      }
                    },
                    "message": {
                      "type": "string",
                      "description": "Message if no stocks found",
                      "example": "No stocks found for this seller"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get all active products",
        "description": "Returns a list of all active products available in the system",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Category ID for filtering"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Product name or SKU for filtering"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/ProductPagination"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "TelegramAuth": []
          }
        ]
      }
    },
    "/api/sales": {
      "post": {
        "tags": ["Sales"],
        "summary": "Create a new sale",
        "description": "Creates a new sale record and updates inventory. Sellers can only sell assigned products from their stock.",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["productId", "quantity", "price"],
                "properties": {
                  "productId": {
                    "type": "string",
                    "description": "Product ID to sell",
                    "example": "64f1a2b3c4d5e6f7g8h9i0j1"
                  },
                  "quantity": {
                    "type": "number",
                    "minimum": 1,
                    "description": "Quantity to sell",
                    "example": 3
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Price per unit",
                    "example": 25000
                  },
                  "customerName": {
                    "type": "string",
                    "description": "Customer name (optional)",
                    "example": "John Doe"
                  },
                  "customerPhone": {
                    "type": "string",
                    "description": "Customer phone number (optional)",
                    "example": "+998901234567"
                  },
                  "notes": {
                    "type": "string",
                    "description": "Additional notes (optional)",
                    "example": "Customer requested express delivery"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sale created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sale": {
                      "$ref": "#/components/schemas/Sale"
                    },
                    "message": {
                      "type": "string",
                      "example": "Sale created successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or insufficient stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Sizda yetarli mahsulot yo'q"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Product not assigned to seller",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "You do not have access to this product"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "tags": ["Sales"],
        "summary": "Get sales history",
        "description": "Returns sales history for the authenticated user with optional date filtering",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date filter (YYYY-MM-DD)",
            "example": "2024-01-01"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date filter (YYYY-MM-DD)",
            "example": "2024-12-31"
          }
        ],
        "responses": {
          "200": {
            "description": "Sales history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sales": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Sale"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/seller/stocks": {
      "get": {
        "tags": ["Seller - Stock Management"],
        "summary": "Get seller's stock inventory",
        "description": "Returns current seller's stock with summary statistics including total products, quantities, and values",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Stock inventory retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stocks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SellerStock"
                      }
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "totalProducts": {
                          "type": "number",
                          "description": "Number of different products",
                          "example": 5
                        },
                        "totalQuantity": {
                          "type": "number",
                          "description": "Total quantity of all products",
                          "example": 150
                        },
                        "totalStockValue": {
                          "type": "number",
                          "description": "Total value of seller's stock",
                          "example": 45000
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Seller access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/seller/stocks/product/{productId}": {
      "get": {
        "tags": ["Seller - Stock Management"],
        "summary": "Get stock for specific product",
        "description": "Returns seller's stock information for a specific product with detailed product information",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID",
            "example": "64f1a2b3c4d5e6f7g8h9i0j1"
          }
        ],
        "responses": {
          "200": {
            "description": "Stock information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stock": {
                      "$ref": "#/components/schemas/SellerStock"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Seller access required"
          },
          "404": {
            "description": "Stock not found for this product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Stock not found for this product"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/seller/products": {
      "get": {
        "tags": ["Seller - Product Management"],
        "summary": "Get seller's assigned products",
        "description": "Returns products assigned to the current seller for selling",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Assigned products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Seller access required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/seller/sales": {
      "get": {
        "tags": ["Seller - Sales Management"],
        "summary": "Get seller's sales history",
        "description": "Returns sales history for the current seller with optional date filtering",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date filter (YYYY-MM-DD)",
            "example": "2024-01-01"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date filter (YYYY-MM-DD)",
            "example": "2024-12-31"
          }
        ],
        "responses": {
          "200": {
            "description": "Sales history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sales": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Sale"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Seller access required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/seller/reports": {
      "get": {
        "tags": ["Seller - Reports"],
        "summary": "Get seller's monthly report",
        "description": "Returns monthly sales report for the current seller with summary statistics and detailed sales data",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "number"
            },
            "description": "Report year (defaults to current year)",
            "example": 2024
          },
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 12
            },
            "description": "Report month (defaults to current month)",
            "example": 12
          }
        ],
        "responses": {
          "200": {
            "description": "Monthly report retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "period": {
                      "type": "object",
                      "properties": {
                        "startDate": {
                          "type": "string",
                          "format": "date-time",
                          "description": "Report period start",
                          "example": "2024-12-01T00:00:00.000Z"
                        },
                        "endDate": {
                          "type": "string",
                          "format": "date-time",
                          "description": "Report period end",
                          "example": "2024-12-31T23:59:59.999Z"
                        }
                      }
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "totalSales": {
                          "type": "number",
                          "description": "Total number of sales",
                          "example": 45
                        },
                        "totalRevenue": {
                          "type": "number",
                          "description": "Total revenue amount",
                          "example": 125000
                        },
                        "totalQuantity": {
                          "type": "number",
                          "description": "Total quantity sold",
                          "example": 300
                        }
                      }
                    },
                    "sales": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Sale"
                      },
                      "description": "Detailed sales data for the period"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Seller access required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}